test(structured): split schema validation coverage
